Avoid turning everyone to es_AR.UTF-8
According to [1], if the system or the user uses an encoding
different from UTF-8, Gnome-Terminal does not start.
Over the 400 labs, many computers are configured with C or
with ISO-8859-1 encoding, so this is a common issue there.

Previously, our script tried to set LANG to es_AR.UTF-8 to
solve this, but that forces this locale to all users.
By making this a conditional check, we make this script
useful to users with different locales, while only forcing
that locale to the ones that are already using es_AR.
Here I am relying on modern distros that set LANG to
whatever.UTF-8 by default.

[1] https://wiki.gnome.org/Apps/Terminal/FAQ#Exit_status_8